+++ common/src/main/java/net/mca/item/BouquetItem.java
@@ -12,7 +12,7 @@ public class BouquetItem extends RelationshipItem {
     }
 
     @Override
-    protected float getHeartsRequired() {
+    protected int getHeartsRequired() {
         return Config.getInstance().bouquetHeartsRequirement;
     }
